Reworked path number parsing, so that they are kept in the integer part of Lua numbers as much as possible.
The default unit for boards is now the picometer. Panelization scripts can offset and merge boards without losing precision (assuming they offset by an integer multiple of the output files epsilon).

A client (grbv for example) can specify unit='mm' in the otions to have the data scaled to millimeters, which may be easier to manipulate. It is however not possible to save such a scaled board yet.